Refactor ModelStub into a nested class in CommandTestUtil
